requirements.txt
ecdsa==0.17.0
fastapi==0.70.0
pydantic==1.8.2
PyJWT==2.3.0
SQLAlchemy==1.4.26
psycopg2-binary
uvicorn